Removed round_ste from max_int, min_int, and max_uint. Moved the functions in <ops> module. Refactored code and doc source. Modified test_ops appropriately
