+++ brevitas/nn/quant_linear.py
@@ -52,7 +52,7 @@ from brevitas.core.scaling import ScalingImplType, SCALING_SCALAR_SHAPE
 from brevitas.core.stats import StatsInputViewShapeImpl
 from brevitas.core.stats import StatsOp
 from brevitas.function.ops import ceil_ste
-from brevitas.function.ops_ste import max_uint
+from brevitas.function.ops import max_uint
 from brevitas.proxy.parameter_quant import WeightQuantProxy, BiasQuantProxy, WeightReg
 from brevitas.config import docstrings
 from .quant_layer import QuantLayer, SCALING_MIN_VAL
