+++ docs/genindex.html
@@ -506,7 +506,7 @@
 </li>
       <li><a href="brevitas.core.html#brevitas.core.stats.StatsOp.MAX_AVE">MAX_AVE (brevitas.core.stats.StatsOp attribute)</a>
 </li>
-      <li><a href="brevitas.function.html#brevitas.function.ops_ste.max_int">max_int() (in module brevitas.function.ops_ste)</a>, <a href="brevitas.function.html#brevitas.function.ops_ste.max_int">[1]</a>
+      <li><a href="brevitas.function.html#brevitas.function.ops.max_int">max_int() (in module brevitas.function.ops)</a>
 </li>
       <li><a href="brevitas.nn.html#brevitas.nn.hadamard_classifier.HadamardClassifier.max_output_bit_width">max_output_bit_width() (brevitas.nn.hadamard_classifier.HadamardClassifier method)</a>
 
@@ -520,7 +520,7 @@
       </ul></li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="brevitas.function.html#brevitas.function.ops_ste.max_uint">max_uint() (in module brevitas.function.ops_ste)</a>, <a href="brevitas.function.html#brevitas.function.ops_ste.max_uint">[1]</a>
+      <li><a href="brevitas.function.html#brevitas.function.ops.max_uint">max_uint() (in module brevitas.function.ops)</a>
 </li>
       <li><a href="brevitas.core.html#brevitas.core.stats.StatsOp.MEAN_LEARN_SIGMA_STD">MEAN_LEARN_SIGMA_STD (brevitas.core.stats.StatsOp attribute)</a>
 </li>
@@ -528,7 +528,7 @@
 </li>
       <li><a href="brevitas.nn.html#brevitas.nn.quant_conv.QuantConv2d.merge_bn_in">merge_bn_in() (brevitas.nn.quant_conv.QuantConv2d method)</a>
 </li>
-      <li><a href="brevitas.function.html#brevitas.function.ops_ste.min_int">min_int() (in module brevitas.function.ops_ste)</a>, <a href="brevitas.function.html#brevitas.function.ops_ste.min_int">[1]</a>
+      <li><a href="brevitas.function.html#brevitas.function.ops.min_int">min_int() (in module brevitas.function.ops)</a>
 </li>
       <li><a href="brevitas.core.html#brevitas.core.bit_width.MsbClampParameterBitWidth">MsbClampParameterBitWidth (class in brevitas.core.bit_width)</a>
 </li>
